+++ src/app/products/products.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { ProductComponent, ProductListComponent } from './components';
 import { SharedModule } from '../shared/shared.module';
@@ -12,7 +11,6 @@ import { SharedModule } from '../shared/shared.module';
     ProductComponent,
   ],
   imports: [
-    CommonModule,
     SharedModule,
   ],
   exports: [